# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mini CRM Platform API",
    "version": "1.0.0",
    "description": "API documentation for the Mini CRM Platform."
  },
  "servers": [
    {
      "url": "http://localhost:5003"
    }
  ],
  "paths": {
    "/api/customers": {
      "post": {
        "summary": "Ingest customer data",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string" },
                  "totalSpend": { "type": "number" },
                  "lastVisit": { "type": "string", "format": "date-time" },
                  "lastOrderDate": { "type": "string", "format": "date-time" },
                  "visitCount": { "type": "integer" }
                },
                "required": ["name", "email"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Customer data queued" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/api/segments/preview": {
      "post": {
        "summary": "Preview segment audience",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "rulesJSON": { "type": "object" } },
                "required": ["rulesJSON"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Audience preview" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid"
      }
    }
  }
}